Removes v0 slot transformers  (#25026)
* Removes v0 slots transformers unit tests


* Removes v0 slot transformers


* Renames v2 transformer file


---------
